[#69] split ActivityAt into separate date and time columns in ActivityRecord
