--- src/main/resources/db/migration/V16__user_activity_record_unique.sql
@@ -1,2 +1,0 @@
-alter table activity_record
-    add constraint uk_user_date unique (user_id, activity_date);
